Move WorkspaceAcl to src/workspaces/common
